Crossover

Let there be documentation on Crossover.

Julia4ML_GA.k_point — Function

k_point(genes1, genes2,rng)

Implements the k point crossover method.

genes1, genes2: Vector{<:Real} containing all genes of a chromosome.

rng: An instance of a random number generator to produce reproducible results.

k: Number of cross over points

Returns child1 and child2

genes1 and genes2 are used to create child1 and child2, which are returned. child1 is a copy of genes1 up to randomly generated crossover point. After that it is a copy of genes2. The same for child2 respectively with parent roles swapped. If more than one crossover point exist this procedure is repeated. If k == 1, this is single point crossover

Julia4ML_GA.partially_mapped — Function

partially_mapped(genes1, genes2,rng)

Implements the partially mapped crossover (PMX) method.

genes1, genes2: Vector{<:Real} containing all genes of a chromosome.

rng: An instance of a random number generator to produce reproducible results.

Returns child1 and child2

genes1 and genes2 are used to create child1 and child2, which are returned.