Examples

Rosenbrock

Julia4ML_GA.solve_rosenbrock — Function

solve_rosenbrock(
    a::Real,
    b::Real;
        starting_point::AbstractVector=Float64[0.,0.],
        iterations=1000, 
        time_limit=NaN, 
        obj_bound=NaN,
        populationSize=100,
        eliteSize=5,
        crossoverRate=0.4,
        mutationRate=0.6,
        selection=roulette_wheel_inv, 
        mutation=gaussian_displacement, 
        crossover=k_point,
        rng=default_rng()
)

Solves the rosenbrock problem. Minimizes the function $(a-x)^2 + b*(y-x^2)^2$. The global minimum $(x,y)$ is at $(a,a^2)$. More information: https://en.wikipedia.org/wiki/Rosenbrock_function

a: First parameter of rosenbrock function.

b: Second parameter of rosenbrock function.

starting_point: Start point + gaussian noise.

max_iterations: Maximum number of iterations in optimisation process.

time_limit: Time in seconds after which the optimization should be terminated.

obj_bound: Threshold on (or after) which the optimization should be terminated.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result.

Rosenbrock Example

julia> import Julia4ML_GA
julia> best_solution = Julia4ML_GA.solve_rosenbrock(4, 100)
2-element Vector{Float64}:
  4.005027433832623
 16.039474997126856

Knapsack

Julia4ML_GA.solve_knapsack — Function

solve_knapsack(
    mass::Vector{<:Real},
    utility::Vector{<:Real},
    maxMass::Real;
        iterations=100, 
        time_limit=NaN, 
        obj_bound=NaN,
        populationSize=50,
        eliteSize=5,
        crossoverRate=0.8,
        mutationRate=0.8,
        selection=roulette_wheel, 
        mutation=bit_inversion, 
        crossover=k_point,
        rng=default_rng()
)

Solves the knapsack problem. More information: https://en.wikipedia.org/wiki/Knapsack_problem

mass: Mass of each item.

utility: Utility of each item.

maxMass: Maximum mass of all items in knapsack.

max_iterations: Maximum number of iterations in optimisation process.

time_limit: Time in seconds after which the optimization should be terminated.

obj_bound: Threshold on (or after) which the optimization should be terminated.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result.

Knapsack Example

julia> import Julia4ML_GA
julia> mass    = [1, 5, 3, 7, 2, 10, 5, 9, 2]
julia> utility = [1, 3, 5, 2, 5,  8, 3, 9, 5]
julia> maxMass = 30
julia> best_solution = Julia4ML_GA.solve_knapsack(mass, utility, maxMass)
julia> print(best_solution)
Bool[1, 0, 1, 0, 1, 1, 0, 1, 1]
julia> sum(best_solution .* utility)
33
julia> sum(best_solution .* mass)
27

Traveling Salesman Problem

Julia4ML_GA.solve_tsp — Function

solve_tsp(
    cost_matrix::Matrix{<:Real};
        iterations=100, 
        time_limit=NaN, 
        obj_bound=NaN,
        populationSize=50,
        eliteSize=5,
        crossoverRate=0.6,
        mutationRate=0.4,
        selection=rank_selection, 
        mutation=displacement, 
        crossover=partially_mapped,
        rng=default_rng()
)

Solves the traveling salesman problem. More information: https://en.wikipedia.org/wiki/Travellingsalesmanproblem

cost_matrix: cost matrix of traveling salesman problem.

max_iterations: Maximum number of iterations in optimisation process.

time_limit: Time in seconds after which the optimization should be terminated.

obj_bound: Threshold on (or after) which the optimization should be terminated.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result.

Traveling Salesman Problem Example

julia> import Julia4ML_GA
julia> cost = [
             0 8 10 7 2;
             8 0 4 5 10;
             10 4 0 3 6;
             7 5 3 0 4;
             2 10 6 4 0
            ]
julia> best_solution, best_solution_cost = Julia4ML_GA.solve_tsp(cost)
([1, 5, 4, 3, 2], 21)

Sudoku

Julia4ML_GA.solve_sudoku — Function

solve_sudoku(
    sudoku::Matrix{<:Integer};
        iterations=10000, 
        time_limit=10, 
        obj_bound=0,
        populationSize=1000,
        eliteSize=5,
        crossoverRate=0.5,
        mutationRate=0.9,
        selection=rank_selection, 
        mutation=sudoku_mutation, 
        crossover=k_point,
        rng=default_rng()
)

Solves the sudoku. More information: https://en.wikipedia.org/wiki/Sudoku

sudoku: sudoku.

max_iterations: Maximum number of iterations in optimisation process.

time_limit: Time in seconds after which the optimization should be terminated.

obj_bound: Threshold on (or after) which the optimization should be terminated.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result.

Sudoku Example

julia> import Julia4ML_GA
julia> sudoku = [
            0 1 0 0 7 5 9 4 0;
            4 0 2 0 0 6 0 0 7;
            0 3 0 9 0 1 0 0 8;
            6 0 0 5 2 0 4 9 0;
            0 5 8 0 0 3 0 2 6;
            1 0 4 0 0 7 0 3 0;
            0 0 9 8 0 0 0 0 2;
            7 0 0 1 6 0 3 0 0;
            2 0 0 0 3 0 6 5 4
            ]
julia> best_solution, errors = Julia4ML_GA.solve_sudoku(sudoku)
julia> print("Number of errors in sudoku: ", errors)
Number of errors in sudoku: 0
julia> display(best_solution)
9×9 Matrix{Int8}:
 8  1  6  2  7  5  9  4  3
 4  9  2  3  8  6  5  1  7
 5  3  7  9  4  1  2  6  8
 6  7  3  5  2  8  4  9  1
 9  5  8  4  1  3  7  2  6
 1  2  4  6  9  7  8  3  5
 3  6  9  8  5  4  1  7  2
 7  4  5  1  6  2  3  8  9
 2  8  1  7  3  9  6  5  4