Initialization

Before the optimization can begin the population has to be initialized. The data type and length of the genome depends on the specific problem to be solved and the representation chosen. During the optimization the size of the population and length of the genome of each individual stays the same.

Template

"""
    template_init_function(population_size::Integer, problem_specific_information::Any, rng::AbstractRNG)

Implements an initialization function.

- `population_size`: Number of populants.
- `problem_specific_information`: Some problem specific information to indicate what possible values each individual can have.
- `rng`: Instance of a random number generator to produce reproducible results.

Returns resulting population.
"""
function template_init_function(population_size::Integer, problem_specific_information::Any, rng::AbstractRNG)
end

Initialization Functions

Julia4ML_GA.init_gaussian — Function

init_gaussian(population_size::Integer, starting_point::AbstractVector, rng::AbstractRNG)

Initializes a start population with populants which have their genome distributed around starting_point.

population_size: Number of populants.

starting_point: The genome of each populant equals this point + gaussian noise.

rng: Instance of a random number generator to produce reproducible results.

Returns population.

Julia4ML_GA.init_uniform_binary_population — Function

init_uniform_binary_population(population_size::Integer, genome_length::Integer, rng::AbstractRNG)

Initializes a start population with populants which have only Bool genes.

population_size: Number of populants.

genome_length: Number of genes of a single populant.

rng: Instance of a random number generator to produce reproducible results.

Returns population.

Julia4ML_GA.init_sudoku_population — Function

init_sudoku_population(population_size::Integer, sudoku::Matrix{Int8}, rng::AbstractRNG)

Initializes a start population with populants which represent a possible solution to the sudoku.

population_size: Number of populants.

sudoku: Length of genes are calculated based on empty (0) places in sudoku.

rng: Instance of a random number generator to produce reproducible results.

Returns population.

Julia4ML_GA.init_tsp_population — Function

init_tsp_population(population_size::Integer, cost_matrix::Matrix{<:Real}, rng::AbstractRNG)

Initializes a start population with populants which represent a possible solution to the tsp provided by the cost_matrix.

population_size: Number of populants.

cost_matrix: Each genome are the shuffled indices, where each index is a stop at that location in the cost_matrix.

rng: Instance of a random number generator to produce reproducible results.

Returns population.