Julia4ML_GA API

Julia4ML_GA.GeneticAlgorithm

Julia4ML_GA.GeneticAlgorithmState

Julia4ML_GA.OptimizationResult

Julia4ML_GA.OptimizationTrace

Julia4ML_GA.Terminator

Julia4ML_GA.append!

Julia4ML_GA.argmin

Julia4ML_GA.bit_inversion

Julia4ML_GA.crossover!

Julia4ML_GA.displacement

Julia4ML_GA.evaluation!

Julia4ML_GA.gaussian_displacement

Julia4ML_GA.get_sub_vector

Julia4ML_GA.k_point

Julia4ML_GA.min

Julia4ML_GA.mutation!

Julia4ML_GA.optimize

Julia4ML_GA.partially_mapped

Julia4ML_GA.rank_selection

Julia4ML_GA.roulette_wheel

Julia4ML_GA.solve_knapsack

Julia4ML_GA.solve_rosenbrock

Julia4ML_GA.solve_sudoku

Julia4ML_GA.solve_tsp

Julia4ML_GA.terminate!

Julia4ML_GA.tournament_selection

Julia4ML_GA.trace

Julia4ML_GA.univariate_displacement

Julia4ML_GA.update_state!

Genetic Algorithm

Julia4ML_GA.GeneticAlgorithm — Type

Contains information and functions to execute an optimization process for a genetic algorithm.

population_size: (Integer) Number of populants to be maintained.

eliteSize: (Integer) Number of populants selected as elite.

crossoverRate: (Float) Probability of crossover for two populants.

mutationRate: (Float) Probability of mutation.

selection: (Function) Function to select populants for next iteration.

muation: (Function) Mutation function.

crossover: (Function) Crossover function.

Constructor:

GeneticAlgorithm(;
    populationSize=50,
    eliteSize=5,
    crossoverRate=0.8,
    mutationRate=0.1,
    selection=roulette_wheel,
    mutation=displacement,
    crossover=k_point
)

Julia4ML_GA.GeneticAlgorithmState — Type

State of a genetic algorithm.

population: (Vector) Current population. Vector conatining elements of same type as starting point for optimization process.

populationFitness: (Function) Fitness function by which the population's fitness is to be evaluated.

Constructor:

GeneticAlgorithmState(population,objective)

Julia4ML_GA.optimize — Function

optimize(starting_point,objective,ga::GeneticAlgorithm;iterations=100,rng=default_rng())

Executes optimization process

starting_point: (Vector) Initial candidate.

ga: (GeneticAlgorithm)

objective: (Function) Fitness function to evaluate population.

iterations: (Real) Maximum number of iterations. Termination condition, defaults to 100.

time_limit: (Real) Time limit in seconds. Termination condition, defaults to NaN.

obj_bound : (Real) Lower bound to objective value. Termination condition, defaults to NaN.

rng: Instance of a random number generator to produce reproducible results. Default is Random.default_rng().

Returns final population's fittest populant.

Population is initialized and build. Then the optimization is executed using the provided fitness function.

Julia4ML_GA.update_state! — Function

update_state!(ga, state, objective, rng)

Updates GeneticAlgorithmState according to provided GeneticAlgorithm instance. Selection, crossover, mutation and evaluation is executed. Equivalent to one iteration of the optimizatrion process.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

state: (GeneticAlgorithmState) GeneticAlgorithmState to proceed from.

objective: (Function) Fitness function to be used.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.crossover! — Function

crossover!(parents,children,selected_individuals,ga,rng)

Control function for crossover.

parents: (Vector) (Sub-)Population to be used to create offspring.

children (Vector) Object to hold the newly created offspring.

selected_individuals: (Vector{Integer}) The for crossover selected populant's indices.

ga: (GeneticAlgorithm) GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.mutation! — Function

mutation!(offspring,ga,rng)

control function for mutation.

population: (Vector) (Sub-)Population to be mutated.

ga: (GeneticAlgorithm) GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.evaluation! — Function

evaluation!(ga,state,objective)

Control function for fitness evaluation.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

state: (GeneticAlgorithmState) GeneticAlgorithmState instance to proceed from.

objective: (Function) Fitness function by which the population is evaluated.

Julia4ML_GA.terminate! — Function

terminate(t::Terminator, state::GeneticAlgorithmState)

Evaluates configured termination criteria.

t: (Terminator) contains termination criteria

state: (GeneticAlgorithmState)

Returns false if algorithm should terminate.

Julia4ML_GA.OptimizationResult — Type

Contains the results of a evolutionary optimization process

minimalPopulant: (T1) Populant that minimizes the objective.

minimalFitness: (T2) Value of the objective for the minimal populant.

trace: (OptimizationTrace) Trace object containing individual values of generations.

Constructor:

OptimizationResult(minimalPopulant,minmalFitness,trace)

Trace

Julia4ML_GA.OptimizationTrace — Type

Contains information and functions to execute an optimization process for a genetic algorithm.

populations: (Vector{Vector{Vector{Real}}}) Population of each iteration.

fitnessValues: (Vector{Vector{Real}}) Fitness values of each generation.

fittestPopulants: (Vector{Vector{Real}}) Fittest Individual of each population.

Constructor:

OptimizationTrace()

Julia4ML_GA.trace — Function

trace(res)

res: (OptimizationResult)

Returns trace of evolutionary algorithm

Helper Functions

Julia4ML_GA.append! — Function

append!(trace,state)

Appends the current state to trace.

trace: (AbstractTrace) Trace struct containing values

state: (AsbtractState) A state of a genetic algorithm.

Julia4ML_GA.argmin — Function

argmin(res)

res: (OptimizationResult)

Returns minimizing gene

Julia4ML_GA.get_sub_vector — Function

get_sub_vector(vec, s, e, allow_wrap)

Returns a part of the vector.

vec: Vector{<:Real}

s: Start Index (including) [1, length]

e: End index (not including) [2, length + 1]

allow_wrap: If true, start can be behind end

Julia4ML_GA.min — Function

min(res)

res: (OptimizationResult)

Returns minimal fitness value