Selection

Let there be documentation on selection.

Julia4ML_GA.roulette_wheel — Function

roulette_wheel(fitness, selection_number, rng)

Implements a simple roulette_wheel. If the sum over positive fitness scores is larger than the sum over negative scores, higher fitness scores are more likely to be selected, otherwise negative values are more likely to be selected. Should only be used if either all fitness values are negative or all fitness values are positive

fitness: (Vector{<:Real}) Vector of fitness values. The higher the absolute fitness, the more likely the corresponding chromosome is selected.

selection_number: (Integer) Indicates how many indices are returned.

rng: Instance of a random number generator to produce reproducible results.

Returns indices of selected populants.

Julia4ML_GA.tournament_selection — Function

tournament_selection(fitness, selection_number, tournament_size, rng)

Implements a simple tournament selection. Selects selection_number candidates. Each candidate is selected by taking the fittest of tournament_size randomly chosen candidates.

fitness: (Vector{<:Real}) Vector of fitness values.

Julia4ML_GA.rank_selection — Function

rank_selection(fitness, selection_number, rng)

Implements rank selection based on roulette_wheel. Can deal with mixed positive and negative values. Selects based on order of fitness values. The amount of the difference between the fitness values is not taken into account.

fitness: (Vector{<:Real}) Vector of fitness values. The lower the fitness, the more likely the corresponding chromosome is selected.

selection_number: (Integer) Indicates how many indices are returned.

rng: Instance of a random number generator to produce reproducible results.

Returns indices of selected populants.