Examples

Rosenbrock

Julia4ML_GA.solve_rosenbrock — Function

solve_rosenbrock(a, b;
iterations, 
time_limit, 
obj_bound,
populationSize,
eliteSize,
crossoverRate,
mutationRate,
selection, 
mutation, 
crossover,
rng)

Solves the rosenbrock problem. Minimizes the function $(a-x)^2 + b*(y-x^2)^2$. The global minimum $(x,y)$ is at $(a,a^2)$. More information: https://en.wikipedia.org/wiki/Rosenbrock_function

a: First parameter of rosenbrock function.

b: Second parameter of rosenbrock function.

max_iterations: Maximum number of iterations in optimisation process. Default is NaN.

time_limit: Time in seconds after which the optimization should be terminated. Default is NaN.

obj_bound: Threshold on (or after) which the optimization should be terminated. Default is NaN.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result

Knapsack

Julia4ML_GA.solve_knapsack — Function

solve_knapsack(mass, utility, maxMass;
iterations, 
time_limit, 
obj_bound,
populationSize,
eliteSize,
crossoverRate,
mutationRate,
selection, 
mutation, 
crossover,
rng)

Solves the knapsack problem. More information: https://en.wikipedia.org/wiki/Knapsack_problem

mass: Mass of each item.

utility: Utility of each item.

maxMass: Maximum mass of all items in knapsack.

max_iterations: Maximum number of iterations in optimisation process. Default is NaN.

time_limit: Time in seconds after which the optimization should be terminated. Default is NaN.

obj_bound: Threshold on (or after) which the optimization should be terminated. Default is NaN.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result

Traveling Salesman Problem

Julia4ML_GA.solve_tsp — Function

solve_tsp(cost_matrix;
iterations, 
time_limit, 
obj_bound,
populationSize,
eliteSize,
crossoverRate,
mutationRate,
selection, 
mutation, 
crossover,
rng)

Solves the traveling salesman problem. More information: https://en.wikipedia.org/wiki/Travellingsalesmanproblem

cost_matrix: cost matrix of traveling salesman problem.

max_iterations: Maximum number of iterations in optimisation process. Default is NaN.

time_limit: Time in seconds after which the optimization should be terminated. Default is NaN.

obj_bound: Threshold on (or after) which the optimization should be terminated. Default is NaN.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result

Sudoku

Julia4ML_GA.solve_sudoku — Function

solve_sudoku(sudoku;
iterations, 
time_limit, 
obj_bound,
populationSize,
eliteSize,
crossoverRate,
mutationRate,
selection, 
mutation, 
crossover,
rng)

Solves the sudoku. More information: https://en.wikipedia.org/wiki/Sudoku

sudoku: sudoku.

max_iterations: Maximum number of iterations in optimisation process. Default is NaN.

time_limit: Time in seconds after which the optimization should be terminated. Default is NaN.

obj_bound: Threshold on (or after) which the optimization should be terminated. Default is NaN.

rng: An instance of a random number generator to produce reproducible results.

population_size: Number of populants to be maintained.

eliteSize: Number of populants selected as elite.

crossoverRate: Probability of crossover for two populants.

mutationRate: Probability of mutation.

selection: Function to select populants for next iteration.

muation: Mutation function.

crossover: Crossover function.

rng: An instance of a random number generator to produce reproducible results.

Returns optimization result