Julia4ML_GA API

Julia4ML_GA.GeneticAlgorithm

Julia4ML_GA.GeneticAlgorithmState

Julia4ML_GA.OptimizationResult

Julia4ML_GA.OptimizationTrace

Julia4ML_GA.Terminator

Julia4ML_GA.append!

Julia4ML_GA.argmin

Julia4ML_GA.bit_inversion

Julia4ML_GA.crossover!

Julia4ML_GA.displacement

Julia4ML_GA.evaluation!

Julia4ML_GA.gaussian_displacement

Julia4ML_GA.get_sub_vector

Julia4ML_GA.init_gaussian

Julia4ML_GA.init_sudoku_population

Julia4ML_GA.init_tsp_population

Julia4ML_GA.init_uniform_binary_population

Julia4ML_GA.k_point

Julia4ML_GA.min

Julia4ML_GA.mutation!

Julia4ML_GA.optimize

Julia4ML_GA.partially_mapped

Julia4ML_GA.rank_selection

Julia4ML_GA.roulette_wheel

Julia4ML_GA.solve_knapsack

Julia4ML_GA.solve_rosenbrock

Julia4ML_GA.solve_sudoku

Julia4ML_GA.solve_tsp

Julia4ML_GA.sudoku_mutation

Julia4ML_GA.terminate!

Julia4ML_GA.tournament_selection

Julia4ML_GA.trace

Julia4ML_GA.univariate_displacement

Julia4ML_GA.update_state!

Genetic Algorithm

Julia4ML_GA.GeneticAlgorithm — Type

Contains information and functions to execute an optimization process for a genetic algorithm.

populationSize::Integer: Number of populants to be maintained.

eliteSize::Integer: Number of populants selected as elite.

mutationRate::Real: Probability of mutation.

crossoverRate::Real: Probability of crossover for two populants.

selection::S: Function to select populants for next iteration.

mutation::M: Mutation function.

crossover::C: Crossover function.

Constructor:

GeneticAlgorithm(;
    populationSize::Integer=50,
    eliteSize::Integer=5,
    crossoverRate::Real=0.5,
    mutationRate::Real=0.5,
    selection::S=roulette_wheel,
    mutation::M=displacement,
    crossover::C=k_point
)

Julia4ML_GA.GeneticAlgorithmState — Type

State of a genetic algorithm.

population::A: Current population. Vector conatining elements of same type as starting point for optimization process.

populationFitness::Vector{<:Real}: Fitness function by which the population's fitness is to be evaluated.

fittest::T: Individual with best fitness value.

Constructor:

GeneticAlgorithmState{T, A}(
    population::A, 
    objective::F
) where {T, A<:AbstractArray, F<:Function}

Julia4ML_GA.optimize — Function

optimize(
    starting_population::AbstractArray,
    objective::Function,
    ga::GeneticAlgorithm;
    iterations::Real = 100, 
    time_limit::Real = NaN, 
    obj_bound::Real = NaN,
    trace_optimization::Bool = false,
    rng::AbstractRNG = default_rng()
)

Executes optimization process. Population is initialized and build. Then the optimization is executed using the provided fitness function.

starting_point: Initial candidate.

objective: Fitness function to evaluate population.

ga: GeneticAlgorithm

iterations: Maximum number of iterations. Termination condition.

time_limit: Time limit in seconds. Termination condition.

obj_bound : Lower bound to objective value. Termination condition.

rng: Instance of a random number generator to produce reproducible results.

Returns final population's fittest populant.

Julia4ML_GA.update_state! — Function

update_state!(ga::GeneticAlgorithm, state::GeneticAlgorithmState, objective::Function, rng::AbstractRNG)

Updates GeneticAlgorithmState according to provided GeneticAlgorithm instance. Selection, crossover, mutation and evaluation is executed. Equivalent to one iteration of the optimizatrion process.

ga: GeneticAlgorithm instance to work on.

state: GeneticAlgorithmState to proceed from.

objective: Fitness function to be used.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.crossover! — Function

crossover!(parents::A, children::A, selected_individuals::Vector{Int}, ga::GeneticAlgorithm, rng::AbstractRNG) where {A<:AbstractArray}

Control function for crossover.

parents: (Sub-)Population to be used to create offspring.

children Object to hold the newly created offspring.

selected_individuals: The for crossover selected populant's indices.

ga: GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.mutation! — Function

mutation!(population::AbstractArray, ga::GeneticAlgorithm, rng::AbstractRNG)

control function for mutation.

population: (Sub-)Population to be mutated.

ga: GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.evaluation! — Function

evaluation!(state::GeneticAlgorithmState, objective::Function)

Control function for fitness evaluation.

state: GeneticAlgorithmState instance to proceed from.

objective: Fitness function by which the population is evaluated.

Julia4ML_GA.terminate! — Function

terminate(t::Terminator, state::GeneticAlgorithmState)

Evaluates configured termination criteria.

t: contains termination criteria

state: GeneticAlgorithmState instance

Returns false if algorithm should terminate.

Julia4ML_GA.OptimizationResult — Type

Contains the results of a evolutionary optimization process

minimalPopulant::T1: Populant that minimizes the objective.

minmalFitness::T2: Value of the objective for the minimal populant.

trace::OptimizationTrace: Trace object containing individual values of generations.

Constructor:

OptimizationResult(minimalPopulant::T1, minimalFitness::T2, trace) where {T1, T2}

Trace

Julia4ML_GA.OptimizationTrace — Type

Contains information and functions to execute an optimization process for a genetic algorithm.

populations::Vector{Vector{Vector{Real}}}: Population of each iteration.

fitnessValues::Vector{Vector{Real}}: Fitness values of each generation.

fittestPopulants::Vector{Vector{Real}}: Fittest Individual of each population.

Constructor:

OptimizationTrace()

Julia4ML_GA.trace — Function

trace(res::OptimizationResult)

res: OptimizationResult

Returns trace of evolutionary algorithm.

Helper Functions

Julia4ML_GA.append! — Function

append!(trace::AbstractTrace, state::AbstractState)

Appends the current state to trace.

trace: Trace struct containing values

state: A state of a genetic algorithm.

Julia4ML_GA.argmin — Function

argmin(res::OptimizationResult)

res: OptimizationResult

Returns minimizing gene.

Julia4ML_GA.get_sub_vector — Function

get_sub_vector(vec::Vector{<:Real}, s::Integer, e::Integer, allow_wrap::Bool=false)

vec: Vector

s: Start Index (including) [1, length]

e: End index (not including) [2, length + 1]

allow_wrap: If true, start can be behind end

Returns a part of the vector.

Julia4ML_GA.min — Function

min(res::OptimizationResult)

res: OptimizationResult

Returns minimal fitness value.