Julia4ML_GA API

Julia4ML_GA.GeneticAlgorithm

Julia4ML_GA.GeneticAlgorithmState

Julia4ML_GA.crossover!

Julia4ML_GA.displacement

Julia4ML_GA.evaluation!

Julia4ML_GA.gaussian_displacement

Julia4ML_GA.initialise_genetic_state

Julia4ML_GA.mutation!

Julia4ML_GA.optimize

Julia4ML_GA.roulette_wheel

Julia4ML_GA.single_point

Julia4ML_GA.update_state!

Julia4ML_GA.GeneticAlgorithm — Type

Contains information and functions to execute an optimization process for a genetic algorithm.

population_size: (Integer) Number of populants to be maintained.

eliteSize: (Integer) Number of populants selected as elite.

crossoverRate: (Float) Probability of crossover for two populants.

mutationRate: (Float) Probability of mutation.

selection: (Function) Function to select populants for next iteration.

muation: (Function) Mutation function.

crossover: (Function) Crossover function.

Constructor:

GeneticAlgorithm(;
    populationSize=50,
    eliteSize=5,
    crossoverRate=0.8,
    mutationRate=0.1,
    selection=roulette_wheel,
    mutation=displacement,
    crossover=single_point
)

Julia4ML_GA.GeneticAlgorithmState — Type

State of a genetic algorithm.

population: (Vector) Current population. Vector conatining elements of same type as starting point for optimization process.

populationFitness: (Function) Fitness function by which the population's fitness is to be evaluated.

Constructor:

GeneticAlgorithmState(population,objective)

Julia4ML_GA.crossover! — Method

crossover!(parents,children,selected_individuals,ga,rng)

Control function for crossover.

parents: (Vector) (Sub-)Population to be used to create offspring.

children (Vector) Object to hold the newly created offspring.

selected_individuals: (Vector{Integer}) The for crossover selected populant's indices.

ga: (GeneticAlgorithm) GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.displacement — Method

displacement(gene,rng)

Implements the single point crossover method.

rng: Instance of a random number generator to produce reproducible results.

gene: (Vector) Vector containing all genes of a chromosome.

Returns resulting gene.

The genes are displaced inside itself. The returned element has the same values as before, but with scrambled genes.

Julia4ML_GA.evaluation! — Method

evaluation!(ga,state,objective)

Control function for fitness evaluation.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

state: (GeneticAlgorithmState) GeneticAlgorithmState instance to proceed from.

objective: (Function) Fitness function by which the population is evaluated.

Julia4ML_GA.gaussian_displacement — Method

gaussian_displacement(gene,rng)

Adds gaussian noise to the gene with $\mathcal{N}(0,1)$.

rng: Instance of a random number generator to produce reproducible results.

gene: (Vector{Float64}) Vector containing all genes of a chromosome.

Returns resulting gene.

Julia4ML_GA.initialise_genetic_state — Method

initialise_genetic_state(starting_point,objective,ga,rng)

Initialises populant's genes.

starting_point: (Vector{Float64})

objective: (Function) Fitness function to be used.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

rng: Instance of a random number generator to produce reproducible results.

Returns GeneticAlgorithmState.

Julia4ML_GA.mutation! — Method

mutation!(offspring,ga,rng)

control function for mutation.

population: (Vector) (Sub-)Population to be mutated.

ga: (GeneticAlgorithm) GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.optimize — Method

optimize(starting_point,objective,ga::GeneticAlgorithm;iterations=100,rng=default_rng())

Executes optimization process

starting_point: (Vector) Initial candidate.

objective: (Function) Fitness function to evaluate population.

iterations: (Integer) Maximum number of iterations.

rng: Instance of a random number generator to produce reproducible results. Default is Random.default_rng().

Returns final population's fittest populant.

Population is initialized and build. Then the optimization is executed using the provided fitness function.

Julia4ML_GA.roulette_wheel — Method

roulette_wheel(fitness, selection_number,rng)

Implements a simple roulette_wheel.

fitness: (Vector{Float64}) Vector of fitness values. The higher, the more likely the corresponding chromosome is selected.

selection_number: (Integer) Indicates how many indices are returned.

rng: Instance of a random number generator to produce reproducible results.

Returns indices of selected populants.

Julia4ML_GA.single_point — Method

single_point(genes1, genes2,rng)

Implements the single point crossover method.

rng: An instance of a random number generator to produce reproducible results.

genes1, genes2, child1 and child2: Vectors containing all genes of a chromosome.

Returns child1 and child2.

genes1 and genes2 are used to create child1 and child2, which are returned. child1 is a copy of genes1 up to randomly generated crossover point. After that it is a copy of genes2. The same for child2 respectively with parent roles swapped.

Julia4ML_GA.update_state! — Method

update_state!(ga, state, objective, rng)

Updates GeneticAlgorithmState according to provided GeneticAlgorithm instance. Selection, crossover, mutation and evaluation is executed. Equivalent to one iteration of the optimizatrion process.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

state: (GeneticAlgorithmState) GeneticAlgorithmState to proceed from.

objective: (Function) Fitness function to be used.

rng: Instance of a random number generator to produce reproducible results.