Julia4ML_GA API

Julia4ML_GA.GeneticAlgorithm

Julia4ML_GA.GeneticAlgorithmState

Julia4ML_GA.OptimizationResult

Julia4ML_GA.OptimizationTrace

Julia4ML_GA.Terminator

Julia4ML_GA.append!

Julia4ML_GA.argmin

Julia4ML_GA.bit_inversion

Julia4ML_GA.crossover!

Julia4ML_GA.displacement

Julia4ML_GA.evaluation!

Julia4ML_GA.gaussian_displacement

Julia4ML_GA.get_sub_vector

Julia4ML_GA.initialise_genetic_state

Julia4ML_GA.k_point

Julia4ML_GA.min

Julia4ML_GA.mutation!

Julia4ML_GA.optimize

Julia4ML_GA.partially_mapped

Julia4ML_GA.rank_selection

Julia4ML_GA.roulette_wheel

Julia4ML_GA.terminate!

Julia4ML_GA.trace

Julia4ML_GA.univariate_displacement

Julia4ML_GA.update_state!

Julia4ML_GA.GeneticAlgorithm — Type

Contains information and functions to execute an optimization process for a genetic algorithm.

population_size: (Integer) Number of populants to be maintained.

eliteSize: (Integer) Number of populants selected as elite.

crossoverRate: (Float) Probability of crossover for two populants.

mutationRate: (Float) Probability of mutation.

selection: (Function) Function to select populants for next iteration.

muation: (Function) Mutation function.

crossover: (Function) Crossover function.

Constructor:

GeneticAlgorithm(;
    populationSize=50,
    eliteSize=5,
    crossoverRate=0.8,
    mutationRate=0.1,
    selection=roulette_wheel,
    mutation=displacement,
    crossover=k_point
)

Julia4ML_GA.GeneticAlgorithmState — Type

State of a genetic algorithm.

population: (Vector) Current population. Vector conatining elements of same type as starting point for optimization process.

populationFitness: (Function) Fitness function by which the population's fitness is to be evaluated.

Constructor:

GeneticAlgorithmState(population,objective)

Julia4ML_GA.OptimizationResult — Type

Contains the results of a evolutionary optimization process

minimalPopulant: (Vector{Vector{Real}}) Populant that minimizes the objective.

minimalFitness: (Vector{Real}) Value of the objective for the minimal populant.

trace: (OptimizationTrace) Trace object containing individual values of generations.

Constructor:

OptimizationResult(minimalPopulant,minmalFitness,trace)

Julia4ML_GA.OptimizationTrace — Type

Contains information and functions to execute an optimization process for a genetic algorithm.

populations: (Vector{Vector{Vector{Real}}}) Population of each iteration.

fitnessValues: (Vector{Vector{Real}}) Fitness values of each generation.

fittestPopulants: (Vector{Vector{Real}}) Fittest Individual of each population.

Constructor:

OptimizationTrace()

Julia4ML_GA.Terminator — Type

Holds information about termination criteria. Starts timer upon creation, if a timelimit is provided.

'max_iterations': (Integer/NaN) Maximum number of iterations in optimisation process. Default is NaN.

'time_limit': (Float/NaN) Time after which the optimization should be terminated. Default is NaN.

'obj_bound': (Float/NaN) Threshold after which the optimization should be terminated. Default is NaN.

Throws - 'ArgumentError' if no termination condition is set. - Warning if no time or iteration limit is provided.

Constructor:

Terminator(;max_iter=NaN, time_limit=NaN, obj_bound=NaN)

Julia4ML_GA.append! — Method

append!(trace,state)

Initialises populant's genes.

trace: (AbstractTrace) Trace struct containing values

state: (AsbtractState) A state of a genetic algorithm.

Julia4ML_GA.argmin — Method

argmin(res)

res: (OptimizationResult)

Returns minimizing gene

Julia4ML_GA.bit_inversion — Method

bit_inversion(gene, rng)

Inverses each bit with probability 1/length(gene).

rng: Instance of a random number generator to produce reproducible results.

gene: (Vector{Bool}) Vector containing all genes of a chromosome.

Returns resulting gene.

!!! Only applicable for problems with true/false genes (knapsack) !!!

Julia4ML_GA.crossover! — Method

crossover!(parents,children,selected_individuals,ga,rng)

Control function for crossover.

parents: (Vector) (Sub-)Population to be used to create offspring.

children (Vector) Object to hold the newly created offspring.

selected_individuals: (Vector{Integer}) The for crossover selected populant's indices.

ga: (GeneticAlgorithm) GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.displacement — Method

displacement(gene, rng)

Implements the displacement method.

gene: ::Vector{<:Real} containing all genes of a chromosome.

rng: Instance of a random number generator to produce reproducible results.

Returns resulting gene.

The genes are displaced inside itself. The returned element has the same values as before, but with scrambled genes.

!!! Only applicable for combinatorial problems (knapsack) !!!

Julia4ML_GA.evaluation! — Method

evaluation!(ga,state,objective)

Control function for fitness evaluation.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

state: (GeneticAlgorithmState) GeneticAlgorithmState instance to proceed from.

objective: (Function) Fitness function by which the population is evaluated.

Julia4ML_GA.gaussian_displacement — Method

gaussian_displacement(gene, rng)

Adds gaussian noise to the gene with $\mathcal{N}(0,1)$.

rng: Instance of a random number generator to produce reproducible results.

gene: (Vector{Float64}) Vector containing all genes of a chromosome.

Returns resulting gene.

!!! Only applicable for numerical problems (rosenbrock) !!!

Julia4ML_GA.get_sub_vector — Function

get_sub_vector(vec, s, e, allow_wrap)

Returns a part of the vector.

vec: Vector{<:Real}

s: Start Index (including) [1, length]

e: End index (not including) [2, length + 1]

allow_wrap: If true, start can be behind end

Julia4ML_GA.initialise_genetic_state — Method

initialise_genetic_state(starting_point,objective,ga,rng)

Initialises populant's genes.

starting_point: (Vector{Float64})

objective: (Function) Fitness function to be used.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

rng: Instance of a random number generator to produce reproducible results.

Returns GeneticAlgorithmState.

Julia4ML_GA.k_point — Function

k_point(genes1, genes2,rng)

Implements the k point crossover method.

genes1, genes2: Vector{<:Real} containing all genes of a chromosome.

rng: An instance of a random number generator to produce reproducible results.

k: Number of cross over points

Returns child1 and child2

genes1 and genes2 are used to create child1 and child2, which are returned. child1 is a copy of genes1 up to randomly generated crossover point. After that it is a copy of genes2. The same for child2 respectively with parent roles swapped. If more than one crossover point exist this procedure is repeated. If k == 1, this is single point crossover

Julia4ML_GA.min — Method

min(res)

res: (OptimizationResult)

Returns minimal fitness value

Julia4ML_GA.mutation! — Method

mutation!(offspring,ga,rng)

control function for mutation.

population: (Vector) (Sub-)Population to be mutated.

ga: (GeneticAlgorithm) GeneticAlgorithm instance the population is part of.

rng: Instance of a random number generator to produce reproducible results.

Julia4ML_GA.optimize — Method

optimize(starting_point,objective,ga::GeneticAlgorithm;iterations=100,rng=default_rng())

Executes optimization process

starting_point: (Vector) Initial candidate.

objective: (Function) Fitness function to evaluate population.

iterations: (Integer) Maximum number of iterations.

rng: Instance of a random number generator to produce reproducible results. Default is Random.default_rng().

Returns final population's fittest populant.

Population is initialized and build. Then the optimization is executed using the provided fitness function.

Julia4ML_GA.partially_mapped — Method

partially_mapped(genes1, genes2,rng)

Implements the partially mapped crossover (PMX) method.

genes1, genes2: Vector{<:Real} containing all genes of a chromosome.

rng: An instance of a random number generator to produce reproducible results.

Returns child1 and child2

genes1 and genes2 are used to create child1 and child2, which are returned.

Julia4ML_GA.rank_selection — Function

rank_selection(fitness, selection_number, rng)

Implements rank selection based on roulette_wheel. Can deal with mixed positive and negative values. Selects based on order of fitness values. The amount of the difference between the fitness values is not taken into account.

fitness: (Vector{<:Real}) Vector of fitness values. The lower the fitness, the more likely the corresponding chromosome is selected.

selection_number: (Integer) Indicates how many indices are returned.

rng: Instance of a random number generator to produce reproducible results.

Returns indices of selected populants.

Julia4ML_GA.roulette_wheel — Method

roulette_wheel(fitness, selection_number, rng)

Implements a simple roulette_wheel. If the sum over positive fitness scores is larger than the sum over negative scores, higher fitness scores are more likely to be selected, otherwise negative values are more likely to be selected. Should only be used if either all fitness values are negative or all fitness values are positive

fitness: (Vector{<:Real}) Vector of fitness values. The higher the absolute fitness, the more likely the corresponding chromosome is selected.

selection_number: (Integer) Indicates how many indices are returned.

rng: Instance of a random number generator to produce reproducible results.

Returns indices of selected populants.

Julia4ML_GA.terminate! — Method

terminate(t::Terminator, state::GeneticAlgorithmState)

Evaluates configured termination criteria.

't': (Terminator) contains termination criteria

'state': (GeneticAlgorithmState)

Returns false if algorithm should terminate.

Julia4ML_GA.trace — Method

trace(res)

res: (OptimizationResult)

Returns trace of evolutionary algorithm

Julia4ML_GA.univariate_displacement — Method

univariate_displacement(gene, rng)

Adds univeriate noise to the gene with $\mathcal{U}(-1,1)$.

rng: Instance of a random number generator to produce reproducible results.

gene: (Vector{Float64}) Vector containing all genes of a chromosome.

Returns resulting gene.

!!! Only applicable for numerical problems (rosenbrock) !!!

Julia4ML_GA.update_state! — Method

update_state!(ga, state, objective, rng)

Updates GeneticAlgorithmState according to provided GeneticAlgorithm instance. Selection, crossover, mutation and evaluation is executed. Equivalent to one iteration of the optimizatrion process.

ga: (GeneticAlgorithm) GeneticAlgorithm instance to work on.

state: (GeneticAlgorithmState) GeneticAlgorithmState to proceed from.

objective: (Function) Fitness function to be used.

rng: Instance of a random number generator to produce reproducible results.